### 2024-06-11 #6871 @editorialbot
[PRE REVIEW]: densify: An R package to prune sparse data frames of typological linguistic data
**Submitting author:** @annagraff (<a href="http://orcid.org/0000-0002-7703-3471">Anna Graff</a>)
**Repository:** https://github.com/annagraff/densify
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.0
