### 2024-07-23 #7024 @editorialbot
[REVIEW]: densify: An R package to prune sparse data frames of typological linguistic data
**Submitting author:** @annagraff (<a href="http://orcid.org/0000-0002-7703-3471">Anna Graff</a>)
**Repository:** https://github.com/annagraff/densify
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.0
