### 2024-07-24 annagraff/densify#7 @elenlefoll
JOSS review 1
Hi!

I am reviewing your JOSS paper on this package (https://github.com/openjournals/joss-reviews/issues/7024) and I am encountering an error on installation:

```
install_github('annagraff/densify', build_vignettes = T)
Downloading GitHub repo annagraff/densify@HEAD
── R CMD build ─────────────────────────────────
✔  checking for file ‘/private/var/folders/_x/ycbk0by91blbvnqkbtv105fw0000gp/T/Rtmp3fjUdu/remotesfb4b4e143e60/annagraff-densify-3e48e73/DESCRIPTION’ ...
─  preparing ‘densify’:
E  creating vignettes (3s)
   --- re-building ‘densify.Rmd’ using rmarkdown
   Error: processing vignette 'densify.Rmd' failed with diagnostics:
   there is no package called ‘bookdown’
   --- failed re-building ‘densify.Rmd’

   SUMMARY: processing the following file failed:
     ‘densify.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error: Failed to install 'densify' from GitHub:
  ! System command 'R' failed
````

I am running:

`> session_info()
─ Session info ───────────────────────────────
 setting  value
 version  R version 4.4.1 (2024-06-14)
 os       macOS Sonoma 14.5
 system   aarch64, darwin20
 ui       RStudio
 language (EN)
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       Europe/Madrid
 date     2024-07-24
 rstudio  2024.04.2+764 Chocolate Cosmos (desktop)
 pandoc   NA`

 Any ideas what is going wrong?
