### 2024-07-24 annagraff/densify#7 @elenlefoll
Also, I wanted to point out that the title of your JOSS paper is potentially misleading: "densify: An R package to prune sparse data frames of typological linguistic data" as, to me, the construction "to prune X of Y" means to get rid of Y. In other words, the title sounds as if {densify} can be used to remove typological linguistic data from sparse data frames.
