### 2024-07-24 annagraff/densify#7 @annagraff
Hello! Regarding your error, have you tried simply installing the bookdown package?

install.packages("bookdown")
remotes::install_github('annagraff/densify', build_vignettes = TRUE)

Regarding the title - happy to put that on the list of things to update!
