### 2024-07-24 annagraff/densify#7 @elenlefoll
Yes, after manually installing bookdown, the installation works.
