+++ DESCRIPTION
@@ -33,6 +33,7 @@ Imports:
     generics (>= 0.1.3),
     cli (>= 3.6.0),
     tibble,
+    bookdown,
     stats
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
