### 2024-07-24 annagraff/densify#7 @elenlefoll
If bookdown is not part of base R, it might make sense to include its installation in the  installation instructions for your package. As a reviewer, I have to answer to tick the box "Installation: Does installation proceed as outlined in the documentation?". For me, strictly speaking, the answer is no.
