### 2024-07-24 annagraff/densify#7 @elenlefoll
The first "demo" command in the paper returns a warning message for me. Is it intended?

```
> example_result <- densify(data = WALS,
+                           taxonomy = glottolog_languoids,
+                           taxon_id = "Glottocode",
+                           density_mean = "log_odds",
+                           min_variability = 3,
+                           limits = list(min_coding_density = 1),
+                           density_mean_weights = list(coding = 1, taxonomy = 1))
Warning message:
In densify(data = WALS, taxonomy = glottolog_languoids, taxon_id = "Glottocode",  :
  ! in `densify()`: no `cols` argument specified,
  using all columns as variables
ℹ use `cols = <tidy column spec>` to silence
  this warning
```
