### 2024-07-24 annagraff/densify#7 @tzakharko
Hi @elenlefoll, thank you for your feedback!

 - The warning is correct (it is a intended to draw the user's attention to default choices), you are however correct that it might be surprising to encounter this warning in demo code. We will update the demo snippets accordingly.

- The code includes automated tests, they are in `tests`.

- My colleagues will probably have a better insight regarding the community guidelines.
