### 2024-07-25 cc0cb00
Added explicit `cols` parameter to the example in the paper in order to silence the warning
