### 2024-07-25 f26eb4d
Added booktabs and ape as suggested dependencies (these are needed to run the tests and build the documentation)
