+++ DESCRIPTION
@@ -33,7 +33,6 @@ Imports:
     generics (>= 0.1.3),
     cli (>= 3.6.0),
     tibble,
-    bookdown,
     stats
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
@@ -44,6 +43,8 @@ Suggests:
     knitr,
     rmarkdown,
     testthat (>= 3.0.0),
-    tidyverse
+    tidyverse,
+    bookdown,
+    ape
 VignetteBuilder: knitr
 Config/testthat/edition: 3
