### 2024-07-26 #7024 @yjunechoe
Just curious - @annagraff do you/coauthors have plans to eventually submit the package to CRAN as well? I have some suggestions which are beyond the standards of JOSS (i.e., not acceptance-blocking) but may be helpful to get the package past the initial submission check by CRAN.
