### 2024-07-26 annagraff/densify#10 @yjunechoe
[JOSS REVIEW] Checklist comments
Dear all,

The submission is in a very good shape. This issue tracks my checklist comments: things that I'd like to see resolved for me to check off all the boxes for my review (I only have a few comments). I will open separate issues for specific comments/suggestions about the code.

## Software Paper

1) I think that the Usage section should be removed from the paper, as example code demonstrating software functionality is typically not within the scope of short JOSS papers (functionality can simply be explained in prose). I would actually like to see the content in the Usage section incorporated into the readme, which is currently very sparse.
1) Related to the above comment about the Usage section, I think the paragraph descriptions from `?densify` and `?densify_result` could be recycled in its place, as they contain succinct summaries of the core features/design of the package.
1) In the paragraph comparing densify with other packages, I would like to know whether densify stands out as being specially tailored for linguistic typology research (I see this domain-specialty as a strength).
1) One small point about rhetoric: I'd prefer the Statement of Need section to start off strong with the compellingly-written second paragraph (starting with "Linguistic typology data is ...") and move the first paragraph (starting with "While the software will..") down, as it reads more like a disclaimer.

## Community Guidelines

I think(?) these are missing (at least I don't see any in `.github/`). For a jumpstart, you could run `usethis::use_tidy_coc()` and `usethis::use_tidy_contributing()` to get templates for "code of conduct" and "contributing" files and edit them accordingly.

---

Ref: https://github.com/openjournals/joss-reviews/issues/7024
