### 2024-07-26 b0483c2
update LICENSE in description (issue #11, point 1)
