### 2024-07-26 faef019
issue #11, point
update tidyverse use to dpylr+ggplot2 in description
