+++ DESCRIPTION
@@ -33,7 +33,8 @@ Imports:
     generics (>= 0.1.3),
     cli (>= 3.6.0),
     tibble,
-    stats
+    stats,
+    dpylr
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.1
@@ -43,7 +44,6 @@ Suggests:
     knitr,
     rmarkdown,
     testthat (>= 3.0.0),
-    tidyverse,
     bookdown,
     ape
 VignetteBuilder: knitr
