### 2024-07-26 5e9b919
issue #11, points 5+6
Update title and description of DESCRIPTION
