+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: densify
-Title: Densifying Input Data Frame
+Title: Densifying sparse data frames of typological linguistic data
 Version: 0.0.0.9000
 Authors@R: 
     c(
@@ -22,7 +22,7 @@ Authors@R:
             role = c("aut")
         )
     )
-Description: This package provides functions to densify data frames with missing values.
+Description: This package provides functions to densify data frames of typological linguistic data containing empty cells to sub-matrices with fewer empty cells. Users can trade off the relative weight given to a series of concerns in this process, including data retention, proportion of non-empty cells and taxonomic diversity of rows. The package therefore allows to densify sparse matrices and to subsample large input matrices in a taxonomy-sensitive procedure.
 License: GNU + file LICENSE
 Depends: R (>= 3.5.0)
 Imports: 
