### 2024-07-26 97f1fa8
Issue #10, points 3+4
Updates to Statement of Need in the paper
