+++ paper/paper.md
@@ -45,14 +45,13 @@ affiliations:
 The R package `densify` provides a procedure to prune input data frames containing empty cells (or cells with values {?} or {NA}) to denser sub-matrices with fewer empty cells. The pruning process trades off a series of variably weighted concerns, including data retention, coding density (proportion of non-empty cells) and taxonomic diversity of rows (representing for example phylogenetic relations). Users can adapt the relative weights given to these concerns through various parameters so that the densification process best fits their needs. As such, the software is useful for several purposes, including the densification of sparse input matrices and the subsampling of large input matrices according to a procedure that is sensitive to taxonomic structure. 
 
 # Statement of Need
-
-While the software will run on any data frame (with rows representing any entities with or without taxonomic structure and columns representing variables), it was primarily designed to prune data frames of linguistic typological data, where the rows are languages (taxa) and the columns typological features (sometimes also called characters, parameters or variables).
-
-Linguistic typological data is increasingly available in large-scale databases, and many analyses that aim at exploring diversity or testing hypotheses rely on such databases. Some of these resources have information for nearly all features across nearly all languages in the database (i.e. they have complete or near-complete coding density), but the dataframe may be too large for certain computationally intensive analyses (e.g., PHOIBLE [@phoible], Grambank [@grambank]). Other databases (e.g., WALS [@wals], AUTOTYP [@AUTOTYP], Lexibank [@Lexibank]) exhibit features that are coded for different sets of languages, resulting in sparse language-feature matrices. Combining data from various databases via language identifiers like glottocodes [@glottolog] usually increases sparsity because the language samples do not match. 
+Linguistic typological data is increasingly available in large-scale databases, and many analyses that aim at exploring diversity or testing hypotheses rely on such databases. Some of these resources have information for nearly all features (sometimes also called characters, parameters or variables) across nearly all languages in the database (i.e. they have complete or near-complete coding density), but the dataframe may be too large for certain computationally intensive analyses (e.g., PHOIBLE [@phoible], Grambank [@grambank]). Other databases (e.g., WALS [@wals], AUTOTYP [@AUTOTYP], Lexibank [@Lexibank]) exhibit features that are coded for different sets of languages, resulting in sparse language-feature matrices. Combining data from various databases via language identifiers like glottocodes [@glottolog] usually increases sparsity because the language samples do not match. 
 
 When datasets are too large or too sparse for computational applications, it can thus be necessary to generate and subsequently operate on a subset of the data represented in a smaller and denser matrix. Thereby, researchers might be particularly interested in maintaining taxonomic diversity in the languages represented, preferentially removing languages belonging to clades represented by many other languages in the sample and penalizing the removal of language isolates or languages which represent small language families.
 
-While certain packages exist to generate sub-matrices from varying input matrices according to principled criteria (e.g., `admmDensestSubmatrix` [@R-admmDensestSubmatrix], which identifies the densest sub-matrix of an input graph of a specified size; or `FSelector` [@R-FSelector] and `varrank` [@R-varrank], which perform attribute subset selection based on various tests and entropy measures to identify the most relevant attributes of a data input), `densify` adds sensitivity to taxonomic structure and generally more flexibility in parameter settings for the user. The pruning algorithm focuses both on the removal of rows and columns and does not require the size of the sub-matrix to be specified a priori.
[…]
