+++ DESCRIPTION
@@ -23,7 +23,7 @@ Authors@R:
         )
     )
 Description: This package provides functions to densify data frames of typological linguistic data containing empty cells to sub-matrices with fewer empty cells. Users can trade off the relative weight given to a series of concerns in this process, including data retention, proportion of non-empty cells and taxonomic diversity of rows. The package therefore allows to densify sparse matrices and to subsample large input matrices in a taxonomy-sensitive procedure.
-License: GNU + file LICENSE
+License: AGPL + file LICENSE
 Depends: R (>= 3.5.0)
 Imports: 
     vctrs (>= 0.6.4),
