### 2024-07-26 7c5d377
Issue #10, points 1+2
Removed examples from usage section, transferring them to the README.
