### 2024-07-26 3b49b1b
Issue #7 (point 2.) and #10 (community guidelines)
Added community guidelines and updated README regarding contributions.
