### 2024-07-26 #7024 @osorensen
> @osorensen I raised this as a checklist comment for the software paper but I now feel a bit uncertain - could you clarify whether demonstrating package design/features via code blocks is allowed for JOSS papers? In the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper), I see:
>
> > Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.
>
> The paper in this submission contains example code, but is not quite _documentation_ of the API, so I'm having trouble evaluating the current case.

Thanks for raising the issue @yjunechoe. Looking at the paper, I feel this is more detailed code than usual JOSS paper. I hence agree with you that this belongs in README and/or vignettes, and that the paper instead should focus on the high-level architecture of the software.
