### 2024-08-21 #7024 @yjunechoe
Just to expedite the process: I've read through the revisions in the draft PR and I'm really happy with all the changes I see, so I've checked off all boxes for my review. I think the submission can move forward to the next step.

One small follow-up suggestion to authors: it'd be nice if the README could also print the output of some of the code being ran (vs. having every line be an assignment to a variable).
