### 2024-08-21 #7024 @osorensen
@annagraff, you don't really have to wait for that, since completing the checklist implies that the review is completed, as far as JOSS is concerned.
