### 2024-08-28 #7024 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.5281/zenodo.7385533 is OK
> - 10.1093/bioinformatics/bty633 is OK
> - 10.5281/zenodo.7976754 is OK
> - 10.5281/zenodo.7836668 is OK
> - 10.5281/zenodo.2677911 is OK
> - 10.5281/zenodo.7740140 is OK
> - 10.5281/zenodo.10804357 is OK
>
> 🟡 SKIP DOIs
>
> - None
>
> ❌ MISSING DOIs
>
> - 10.32614/cran.package.admmdensestsubmatrix may be a valid DOI for title: admmDensestSubmatrix: Alternating Direction Method...
> - 10.32614/cran.package.fselector may be a valid DOI for title: FSelector: Selecting Attributes
> - 10.32614/cran.package.varrank may be a valid DOI for title: varrank: Heuristics Tools Based on Mutual Informat...
>
> ❌ INVALID DOIs
>
> - None
> ```

@annagraff, I've now read through the paper, and everything looks good. However, could you please add the suggested DOIs to the R packages that you are citing, and then let me know when this is done?
