### 2024-08-28 711299c
Added DOIs
Added 3 DOIs to paper
