+++ paper/sources.bib
@@ -14,7 +14,8 @@ Dubmatrix Problem},
     author = {Brendan Ames and Polina Bombina},
     year = {2019},
     note = {R package version 0.1.0},
-    url = {https://CRAN.R-project.org/package=admmDensestSubmatrix},
+    url = {https://doi.org/10.32614/cran.package.admmdensestsubmatrix},
+	doi = {10.32614/cran.package.admmdensestsubmatrix}
   }
 
   @Manual{R-FSelector,
@@ -22,7 +23,8 @@ Dubmatrix Problem},
     author = {Piotr Romanski and Lars Kotthoff and Patrick Schratz},
     year = {2023},
     note = {R package version 0.34},
-    url = {https://CRAN.R-project.org/package=FSelector},
+    url = {https://doi.org/10.32614/cran.package.fselector},
+	doi = {10.32614/cran.package.fselector}
   }
 
   @Article{R-ape,
@@ -103,5 +105,6 @@ Dubmatrix Problem},
     author = {Gilles Kratzer and Reinhard Furrer},
     year = {2022},
     note = {R package version 0.5},
-    url = {https://CRAN.R-project.org/package=varrank},
+    url = {https://doi.org/10.32614/cran.package.varrank},
+	doi = {10.32614/cran.package.varrank}
   }
