### 2024-08-28 #7024 @annagraff
@osorensen <https://github.com/osorensen>:
- The version tag is 1.0.0.
- The DOI of the archival deposit on Zenodo is the following: 10.5281/zenodo.13384828

> On 28 Aug 2024, at 16:04, Øystein Sørensen ***@***.***> wrote:
>
>
> @annagraff <https://github.com/annagraff>, at this point could you:
>
> Make a tagged release of your software, and list the version tag of the archived version here.
> Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> Please list the DOI of the archived version here.
> I can then move forward with recommending acceptance.
>
