### 2024-08-28 #7024 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
